bitfield.hpp and bitfield_impl.hpp fusioned
