+++ hdl/common/include/util/memory_mapping/bitfield.hpp
@@ -19,40 +19,68 @@ namespace MemoryMapping {
   public:
     //! \brief Constructor
     //! \param reg : The Register which this Field belongs to
-    constexpr BitField(const Register<RegType, Group>& reg);
+    constexpr BitField(const Register<RegType, Group>& reg)
+      : REGISTER(reg) {
+    }
 
     //! \brief Cast operator
-    operator bool(void) const;
+    operator bool(void) const {
+      return (REGISTER >> BITNUM) & 1;
+    }
 
     // Assignment
 
     //! \brief Assignment operator for a Config object.
-    const BitField& operator=(const Config<RegType, Group, (1<<BITNUM)>& cfg) const;
+    const BitField& operator=(const Config<RegType, Group, (1<<BITNUM)>& cfg) const {
+      REGISTER = (REGISTER & ~(1 << BITNUM)) | cfg.VALUE;
+      return *this;
+    }
 
     //! \brief Assignment operator for an other BitField.
     /*!
      * This is a simplified version for the case where the bitfields have the same BITNUM.
      */
-    const BitField& operator=(const BitField<RegType, Group, BITNUM>& bf) const;
+    const BitField& operator=(const BitField<RegType, Group, BITNUM>& bf) const {
+      REGISTER = (REGISTER & ~(1 << BITNUM)) | (bf.REGISTER & (1 << BITNUM));
+      return *this;
+    }
 
     //! \brief Assignment operator for an other BitField.
     /*!
      * Unlike the other classes, BitFields can be used with any other BitField,
      * because a boolean is a boolean !
      */
-    template<typename OtherRegType, typename OtherGroup, int OTHER_BITNUM> const BitField& operator=(const BitField<OtherRegType, OtherGroup, OTHER_BITNUM>& bf) const;
+    template<typename OtherRegType, typename OtherGroup, int OTHER_BITNUM> const BitField& operator=(const BitField<OtherRegType, OtherGroup, OTHER_BITNUM>& bf) const {
+      REGISTER = (REGISTER & ~(1 << BITNUM)) | (((bf.REGISTER >> OTHER_BITNUM) & 1) << BITNUM);
+      return *this;
+    }
     
     //! \brief Assignment operator for a bool value.
-    const BitField& operator=(const bool val) const;
+    const BitField& operator=(const bool val) const {
+      if(val) {
+	REGISTER |= (1 << BITNUM);
+      }
+      else {
+	REGISTER &= (RegType)(~(1 << BITNUM));
+      }
+      return *this;
+    }
 
     // Logic
-    #define OP &&
-    #include "bitfield_decl_logic.hpp"
-    #undef OP
+#define MACRO_DEFINE_LOGIC(op)						\
+    inline bool operator op(const Config<RegType, Group, (1<<BITNUM)>& cfg) const { \
+      return ((bool)*this) op cfg;					\
+    }									\
+    template<typename OtherRegType, typename OtherGroup, int OTHER_BITNUM> \
+    inline bool operator op(const BitField<OtherRegType, OtherGroup, OTHER_BITNUM>& bf) const { \
+      return ((bool)*this) op ((bool)bf);				\
+    }									\
+    inline bool operator op(const bool val) const {			\
+      return ((bool)*this) op val;					\
+    }
 
-    #define OP ||
-    #include "bitfield_decl_logic.hpp"
-    #undef OP
+    MACRO_DEFINE_LOGIC(&&);
+    MACRO_DEFINE_LOGIC(||);
 
   };
